+++ history.md
@@ -1,6 +1,8 @@
 ### History
 
 v1.1.0dev
+- Check the selected layer and deactivate plugin if the layer is not compatible with InSAR Explorer.
+- Add support for polygon selection for vector data.
 - Change the reference point for the map on the fly.
 - Enhance labeling of the map plot
 - Enhance curve fitting by a prior normalizing of dates.